Move temp device type to sensor
Move temp device type to sensor
