+++ circuitmatter/__main__.py
@@ -12,7 +12,7 @@ import time
 import circuitmatter as cm
 
 from circuitmatter.device_types.lighting import on_off
-from circuitmatter.device_types.measurement import temperature_sensor
+from circuitmatter.device_types.sensor import temperature_sensor
 
 
 class ReplaySocket:
